+++ lib/Utils/routes.dart
@@ -1,4 +1,5 @@
 class MyRoutes {
   static String loginRoute = "/LoginPage";
   static String homeRoute = "/HomePage";
+  static String homeDetailsRoute = "/detail";
 }
