+++ lib/pages/home_detail_page.dart
@@ -1,0 +1,75 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:thirty_days_of_flutter/main.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+import 'package:thirty_days_of_flutter/models/catalog.dart';
+
+class HomeDetailsPage extends StatelessWidget {
+  const HomeDetailsPage({
+    Key? key,
+    required this.catalog,
+  })  : assert(catalog != null),
+        super(key: key);
+
+  final Item catalog;
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(),
+        backgroundColor: MyApp.creamColor,
+        bottomNavigationBar: Container(
+          color: Colors.white,
+          child: ButtonBar(
+            alignment: MainAxisAlignment.spaceBetween,
+            buttonPadding: EdgeInsets.zero,
+            children: [
+              "\$${catalog.price}".text.bold.xl4.red800.make(),
+              ElevatedButton(
+                      onPressed: () {},
+                      style: ButtonStyle(
+                          backgroundColor:
+                              MaterialStateProperty.all(MyApp.darkBluesh),
+                          shape:
+                              MaterialStateProperty.all(const StadiumBorder())),
+                      child: "Buy".text.color(Colors.white).make())
+                  .wh(100, 50),
+            ],
+          ).p32(),
+        ),
+        body: SafeArea(
+          bottom: false,
+          child: Column(
+            children: [
+              Hero(
+                tag: Key(catalog.id.toString()),
+                child: Image.network(catalog.image),
+              ).h32(context),
+              Expanded(
+                  child: VxArc(
+                height: 30.0,
+                arcType: VxArcType.convey,
+                edge: VxEdge.top,
+                child: Container(
+                    color: Colors.white,
+                    width: context.screenWidth,
+                    child: Column(
+                      children: [
+                        catalog.name.text.xl4
+                            .color(MyApp.darkBluesh)
+                            .bold
+                            .make(),
+                        catalog.desc.text.xl
+                            .textStyle(context.captionStyle)
+                            .make(),
+                        10.heightBox,
+                      ],
+                    ).py(64)),
+              ))
+            ],
+          ),
+        ));
+  }
+}
