+++ lib/pages/home_page.dart
@@ -3,6 +3,8 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:thirty_days_of_flutter/widgets/home_Widgets/catalog_header.dart';
+import 'package:thirty_days_of_flutter/widgets/home_Widgets/catalog_list.dart';
 import 'package:velocity_x/velocity_x.dart';
 
 import 'package:thirty_days_of_flutter/main.dart';
@@ -52,110 +54,12 @@ class _HomePageState extends State<HomePage> {
               children: [
                 const CatalogHeader(),
                 if (CatalogModel.items != null && CatalogModel.items.isNotEmpty)
-                  const CatalogList().expand()
+                  const CatalogList().py16().expand()
                 else
-                  const Center(child: CircularProgressIndicator())
+                  const CircularProgressIndicator().centered().expand()
               ],
             ),
           ),
         ));
-  }
-}
-
-class CatalogHeader extends StatelessWidget {
-  const CatalogHeader({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        "Pradip's Tech".text.xl5.bold.color(MyApp.darkBluesh).make(),
-        "Trending Products".text.xl2.make()
-      ],
-    );
-  }
-}
-
-class CatalogList extends StatelessWidget {
-  const CatalogList({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView.builder(
-      shrinkWrap: true,
-      itemCount: CatalogModel.items.length,
-      itemBuilder: ((context, index) {
-        final catalog = CatalogModel.items[index];
-        return CatalogItem(catalog: catalog);
-      }),
-    );
-  }
-}
-
-class CatalogItem extends StatelessWidget {
-  final Item catalog;
-
-  const CatalogItem({
-    Key? key,
-    required this.catalog,
-  })  : assert(catalog != null),
-        super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return VxBox(
-        child: Row(
-      children: [
-        CatalogImage(image: catalog.image),
-        Expanded(
-            child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            catalog.name.text.lg.color(MyApp.darkBluesh).bold.make(),
-            catalog.desc.text.textStyle(context.captionStyle).make(),
-            10.heightBox,
-            ButtonBar(
-              alignment: MainAxisAlignment.spaceBetween,
-              buttonPadding: EdgeInsets.zero,
-              children: [
-                "\$${catalog.price}".text.bold.xl.make(),
-                ElevatedButton(
-                    onPressed: () {},
-                    style: ButtonStyle(
-                        backgroundColor:
-                            MaterialStateProperty.all(MyApp.darkBluesh),
-                        shape:
-                            MaterialStateProperty.all(const StadiumBorder())),
-                    child: "Buy".text.color(Colors.white).make())
-              ],
-            ).pOnly(right: 8.0)
-          ],
-        ))
-      ],
-    )).white.rounded.square(150).make().py16();
-  }
-}
-
-class CatalogImage extends StatelessWidget {
-  const CatalogImage({
-    Key? key,
-    required this.image,
-  })  : assert(image != null),
-        super(key: key);
-
-  final String image;
-
-  @override
-  Widget build(BuildContext context) {
-    return Image.network(image)
-        .box
-        .rounded
-        .p8
-        .color(MyApp.creamColor)
-        .make()
-        .p16()
-        .w40(context);
   }
 }
