+++ lib/widgets/home_Widgets/catalog_header.dart
@@ -1,0 +1,19 @@
+//CatalogHeader
+import 'package:flutter/material.dart';
+import 'package:thirty_days_of_flutter/main.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class CatalogHeader extends StatelessWidget {
+  const CatalogHeader({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        "Pradip's Tech".text.xl5.bold.color(MyApp.darkBluesh).make(),
+        "Trending Products".text.xl2.make()
+      ],
+    );
+  }
+}
