+++ lib/widgets/home_Widgets/catalog_item.dart
@@ -1,0 +1,54 @@
+//CatalogItem
+import 'package:flutter/material.dart';
+import 'package:thirty_days_of_flutter/main.dart';
+import 'package:thirty_days_of_flutter/models/catalog.dart';
+import 'package:thirty_days_of_flutter/widgets/home_Widgets/catalog_image.dart';
+import 'package:thirty_days_of_flutter/pages/home_page.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class CatalogItem extends StatelessWidget {
+  final Item catalog;
+
+  const CatalogItem({
+    Key? key,
+    required this.catalog,
+  })  : assert(catalog != null),
+        super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return VxBox(
+        child: Row(
+      children: [
+        Hero(
+            tag: Key(catalog.id.toString()),
+            child: CatalogImage(image: catalog.image)),
+        Expanded(
+            child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            catalog.name.text.lg.color(MyApp.darkBluesh).bold.make(),
+            catalog.desc.text.textStyle(context.captionStyle).make(),
+            10.heightBox,
+            ButtonBar(
+              alignment: MainAxisAlignment.spaceBetween,
+              buttonPadding: EdgeInsets.zero,
+              children: [
+                "\$${catalog.price}".text.bold.xl.make(),
+                ElevatedButton(
+                    onPressed: () {},
+                    style: ButtonStyle(
+                        backgroundColor:
+                            MaterialStateProperty.all(MyApp.darkBluesh),
+                        shape:
+                            MaterialStateProperty.all(const StadiumBorder())),
+                    child: "Buy".text.color(Colors.white).make())
+              ],
+            ).pOnly(right: 8.0)
+          ],
+        ))
+      ],
+    )).white.rounded.square(150).make().py16();
+  }
+}
